[REFACTORING] Split IMAPServerSSLTest
