+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerTest.java
@@ -33,7 +33,6 @@ import static org.mockito.Mockito.spy;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
-import java.net.Socket;
 import java.nio.ByteBuffer;
 import java.nio.channels.SocketChannel;
 import java.nio.charset.StandardCharsets;
@@ -46,8 +45,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Predicate;
 import java.util.stream.IntStream;
 
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
 
 import jakarta.mail.Flags;
@@ -253,54 +250,6 @@ class IMAPServerTest {
 
 
 
-    @Nested
-    class SSL {
-        IMAPServer imapServer;
-        private MailboxSession mailboxSession;
-        private MessageManager inbox;
-        private Socket clientConnection;
-
-        @BeforeEach
-        void beforeEach() throws Exception {
-            imapServer = createImapServer("imapServerSSL.xml");
-            int port = imapServer.getListenAddresses().get(0).getPort();
-            mailboxSession = memoryIntegrationResources.getMailboxManager().createSystemSession(USER);
-            memoryIntegrationResources.getMailboxManager()
-                .createMailbox(MailboxPath.inbox(USER), mailboxSession);
-            inbox = memoryIntegrationResources.getMailboxManager().getMailbox(MailboxPath.inbox(USER), mailboxSession);
-
-            SSLContext ctx = SSLContext.getInstance("TLS");
-            ctx.init(null, new TrustManager[]{new BlindTrustManager()}, null);
-            clientConnection = ctx.getSocketFactory().createSocket();
-            clientConnection.connect(new InetSocketAddress(LOCALHOST_IP, port));
-            byte[] buffer = new byte[8193];
-            clientConnection.getInputStream().read(buffer);
-        }
-
-        @AfterEach
-        void tearDown() throws Exception {
-            clientConnection.close();
-            imapServer.destroy();
-        }
-
-        @Test
-        void startTlsCapabilityShouldFailWhenSSLSocket() throws Exception {
-            clientConnection.getOutputStream().write("a0 STARTTLS\r\n".getBytes(StandardCharsets.UTF_8));
-            assertThat(readString(clientConnection)).startsWith("a0 BAD STARTTLS failed. Unknown command.");
-        }
-
-        @Test
-        void startTlsCapabilityShouldNotBeAdvertisedWhenSSLSocket() throws Exception {
-            clientConnection.getOutputStream().write("a0 CAPABILITY\r\n".getBytes(StandardCharsets.UTF_8));
-            assertThat(readString(clientConnection)).doesNotContain("STARTTLS");
-        }
-
-        private String readString(Socket channel) throws IOException {
-            byte[] buffer = new byte[8193];
-            int read = channel.getInputStream().read(buffer);
-            return new String(buffer, 0, read, StandardCharsets.US_ASCII);
-        }
-    }
 
     @Nested
     class IdleSSLCompress {
